refactor: collected all exceptions in one central file
